add user role filtering to manage-users, add asc/desc order by for username column in manage-users

+++ public_html/Project/manage-users.php
@@ -20,12 +20,12 @@
 
         switch ($action) {
             case 'create-user':
-                $username  = $_POST['username'] ?? '';
+                $username = $_POST['username'] ?? '';
                 $firstname = $_POST['firstname'] ?? '';
-                $lastname  = $_POST['lastname'] ?? '';
-                $email     = $_POST['email'] ?? '';
-                $password  = $_POST['password'] ?? '';
-                $role      = $_POST['role'] ?? '';
+                $lastname = $_POST['lastname'] ?? '';
+                $email = $_POST['email'] ?? '';
+                $password = $_POST['password'] ?? '';
+                $role = $_POST['role'] ?? '';
                 createUser($username, $firstname, $lastname, $email, $password, $role);
                 updateLastSeen($_SESSION['user_id']);
                 break;
@@ -37,6 +37,10 @@
                 break;
             case 'delete-user':
                 $userId = (int)$_POST['user_id'];
+                if ($userId == $_SESSION['user_id']) {
+                    echo '<script> alert("Deleting yourself is a bad idea..") </script>';
+                    break;
+                }
                 deleteUser($userId);
                 updateLastSeen($_SESSION['user_id']);
                 break;
@@ -46,17 +50,38 @@
         }
     }
 
-    $page = $_GET['page'] ?? 1;
-    $limit = $_GET['limit'] ?? 24;
+    $page = (int)$_GET['page'] ?? 1;
+    $limit = (int)$_GET['limit'] ?? 24;
+    $abc = $_GET['abc'] ?? 'ASC';
+    $roleFilter = $_GET['role_filter'] ?? NULL;
+
+    $allowedAbc = ['ASC', 'DESC'];
+    if (!in_array($abc, $allowedAbc, TRUE)) {
+        $abc = 'ASC';
+    }
 
     $allowedLimits = [12, 24, 48];
-
-    if (!in_array($limit, $allowedLimits)) {
+    if (!in_array($limit, $allowedLimits, TRUE)) {
         $limit = 24;
     }
 
-    $userCount = getUserCount();
-    $totalPages =  ceil($userCount / $limit);
+    if ($roleFilter) {
+        $allowedRoles = ['administrator', 'responder', 'reporter'];
+        $iterations = count($roleFilter);
+        if ($iterations > 3) {
+            $roleFilter = NULL;
+        } else {
+            for ($i = 0; $i < $iterations; $i++) {
+                if (!in_array($roleFilter[$i], $allowedRoles)) {
+                    $roleFilter = NULL;
+                    break;
+                }
+            }
+        }
+    }
+
+    $userCount = getUserCount($roleFilter);
+    $totalPages =  max(1, ceil($userCount / $limit));
     
     if ($page > $totalPages) {
         $page = $totalPages;
@@ -66,8 +91,21 @@
     }
         
     $offset = ($page - 1) * $limit;
-    $users = getUsers($limit, $offset);
+    $users = getUsers($limit, $offset, $abc, $roleFilter);
     $now = time();
+
+    function roleQuery($roles) {
+        if (empty($roles) || !is_array($roles)) {
+            return '';
+        }
+
+        $query = '';
+        foreach ($roles as $role) {
+            $query .= '&role_filter[]=' . urlencode($role);
+        }
+
+        return $query;
+    }
 ?>
 
 <?php require_once 'includes/header.php' ?>
@@ -87,13 +125,13 @@
                             <?php else: ?>
                             <div class='user-card-header offline'>
                             <?php endif; ?>
-                                <h2><?= htmlspecialchars($user['username']) ?> <?php if($user['user_id'] === $_SESSION['user_id']) {echo '(me)';} ?></h2>
+                                <h2><?= htmlspecialchars($user['username']) ?> <?php if($user['user_id'] == $_SESSION['user_id']) {echo '(me)';} ?></h2>
                                 <p><?= htmlspecialchars($user['user_role']) ?></p>
                             </div>
                             <div class='user-card-info'>
-                                <p>Name: <?= $user['full_name'] ?></p>
-                                <p>E-mail: <?= $user['user_email'] ?></p>
-                                <p>ID: <?= $user['user_id'] ?></p>
+                                <p>Name: <?= htmlspecialchars($user['full_name']) ?></p>
+                                <p>E-mail: <?= htmlspecialchars($user['user_email']) ?></p>
+                                <p>ID: <?= htmlspecialchars($user['user_id']) ?></p>
                             </div>
                             <div class='user-card-options'>
                                 <form method='POST'>
@@ -117,24 +155,49 @@
                 </div>
                 <div id='pagination-div'>
                     <nav class='pagination'>
-                        <a href='?page=1&limit=<?= $limit ?>' id='back-to-start'><<</a>
-                        <a href='?page=<?= max(1, $page - 1); ?>&limit=<?= $limit ?>' id='next'><</a>
+                        <a href='?page=1&limit=<?= $limit ?>&abc=<?= $abc ?><?= roleQuery($roleFilter) ?>' id='back-to-start'><<</a>
+                        <a href='?page=<?= max(1, $page - 1); ?>&limit=<?= $limit ?>&abc=<?= $abc ?><?= roleQuery($roleFilter) ?>' id='next'><</a>
                         <?php for ($i = 1; $i <= $totalPages; $i++): ?>
-                            <a href='?page=<?= $i ?>&limit=<?= $limit ?>' class='<?= $i == $page ? "active" : "" ?>'><?= $i ?></a>
+                            <a href='?page=<?= $i ?>&limit=<?= $limit ?>&abc=<?= $abc ?><?= roleQuery($roleFilter) ?>' class='<?= $i == $page ? "active" : "" ?>'><?= $i ?></a>
                         <?php endfor; ?>
-                        <a href='?page=<?= min($totalPages, $page + 1); ?>&limit=<?= $limit ?>' id='previous'>></a>
-                        <a href='?page=<?= $totalPages ?>&limit=<?= $limit ?>' id='skip-to-end'>>></a>
+                        <a href='?page=<?= min($totalPages, $page + 1); ?>&limit=<?= $limit ?>&abc=<?= $abc ?><?= roleQuery($roleFilter) ?>' id='previous'>></a>
+                        <a href='?page=<?= $totalPages ?>&limit=<?= $limit ?>&abc=<?= $abc ?><?= roleQuery($roleFilter) ?>' id='skip-to-end'>>></a>
                     </nav>
                 </div>
             </div>
             <div id='options-panel'>
-                <form method='GET' id='limit-form'>
+                <form method='GET' id='options-form'>
                     <label for='limit'>Users per page:</label>
                     <select name='limit' id='limit' onchange='this.form.submit()'>
                         <option value='12' <?= $limit == 12 ? 'selected' : '' ?>>12</option>
                         <option value='24' <?= $limit == 24 ? 'selected' : '' ?>>24</option>
                         <option value='48' <?= $limit == 48 ? 'selected' : '' ?>>48</option>
                     </select>
+
+                    <input type='hidden' name='abc' id='abc-value' value="<?= $abc ?>">
+                    <button type='button' id='abc-toggle'>
+                        abc <?= $abc ?>
+                    </button>
+                    <div id='role-filtering'>
+                        <label>
+                            <input type='checkbox' name='role_filter[]' value='administrator'
+                                <?= in_array('administrator', $_GET['role_filter'] ?? []) ? 'checked' : '' ?>>
+                            Administrator
+                        </label>
+
+                        <label>
+                            <input type='checkbox' name='role_filter[]' value='responder'
+                                <?= in_array('responder', $_GET['role_filter'] ?? []) ? 'checked' : '' ?>>
+                            Responder
+                        </label>
+
+                        <label>
+                            <input type='checkbox' name='role_filter[]' value='reporter'
+                                <?= in_array('reporter', $_GET['role_filter'] ?? []) ? 'checked' : '' ?>>
+                            Reporter
+                        </label>                    
+                    </div>
+                    
                 </form>
                 <button type='button' id='open-create-user-panel-btn'>Add user</button>
             </div>
